config/src/main/java/com/marketplace/config/ConfigLoader.java
package com.marketplace.config;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.nio.charset.Charset;

public class ConfigLoader {

    public static <T> T loadClasspathResource(String resource, Class<T> tClass) throws IOException {
        InputStream resourceAsStream = Thread.currentThread().getContextClassLoader().getResourceAsStream(resource);
        return ConfigModule.objectMapper()
                .readValue(resourceAsStream, tClass);
    }

    public static String resourceAsString(String resource) throws IOException {
        InputStream resourceAsStream = Thread.currentThread().getContextClassLoader().getResourceAsStream(resource);
       return convert(resourceAsStream, Charset.defaultCharset());
    }

    public static String convert(InputStream inputStream, Charset charset) throws IOException {
        StringBuilder stringBuilder = new StringBuilder();
        String line;
        try (BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream, charset))) {
            while ((line = bufferedReader.readLine()) != null) {
                stringBuilder.append(line);
            }
        }

        return stringBuilder.toString();
    }
}
